+++ src/main/java/Lesson10/App/TurnOffOn.java
@@ -1,0 +1,16 @@
+package Lesson10.App;
+
+public interface TurnOffOn {
+
+
+
+    static boolean appOn(Object o){
+        System.out.println("If turn on - true, turn on - false:");
+        return true;
+    }
+
+     static boolean appOff(Object o){
+         System.out.println("If turn on - true, turn on - false:");
+        return false;
+    }
+}
